Fix premature unmounting during switch to sysroot
During boot, initrd-switch-root.target is isolated, meaning that all
services started prior to this target which are not dependencies of it
are being stopped. Set IgnoreOnIsolate=true to avoid stopping the mount
service provided by this module, and thus avoid prematurely unmounting
the respective partition before we actually get a chance to escape
initramfs.

Also add a dependancy on the service to initrd-switch-root.target. Note
that this is not strictly necessary with the IgnoreOnIsolate flag set.

NB: The BindsTo option should make sure that the services are stopped
when the devices are brought down.